Removed an unnecessarily controversial line
It's misleading and imho downright false to claim that C# was usually the best choice for companies.
It should be obvious that the time spent learning GDscript is well invested. If that wasn't the case then why does almost any other engine use a custom scripting language? 

I am not saying that there aren't companies where using C# works well or where it might even be the better choice for some reason, but this line is unnecessary at this place and will only lead to false impressions. If you are in a company that intends to use C# you won't need this information. If you don't know about C# yet you are probably better off using GDscript.